Pulled pricing out into it's own folder and typed actions/reducers/epics
