+++ src/client/src/combineReducers.ts
@@ -3,7 +3,7 @@ import compositeStatusServiceReducer from './compositeStatusServiceOperations'
 import { connectionStatusReducer } from './connectionStatusOperations'
 import { currencyPairReducer } from './currencyPairsOperations'
 import { enviromentReducer } from './enviromentReducer'
-import { pricingServiceReducer } from './pricingOperations'
+import { pricingServiceReducer } from './operations/pricing'
 import { analyticsReducer } from './ui/analytics'
 import { blotterReducer } from './ui/blotter'
 import { regionsReducer } from './ui/common/regions/regionsOperations'
