+++ src/client/src/configureStore.ts
@@ -8,7 +8,7 @@ import { compositeStatusServiceEpic } from './compositeStatusServiceOperations'
 import { connectionStatusEpicsCreator } from './connectionStatusOperations'
 import { linkEpic } from './linkEpic'
 import { openfinEpic } from './openfinEpics'
-import { pricingServiceEpic } from './pricingOperations'
+import { pricingServiceEpic } from './operations/pricing'
 import { referenceServiceEpic } from './referenceDataOperations'
 import { analyticsServiceEpic } from './ui/analytics'
 import { blotterEpic } from './ui/blotter/'
