+++ src/client/src/operations/pricing/index.ts
@@ -1,0 +1,3 @@
+export { PricingActions, ACTION_TYPES } from './actions'
+export { pricingServiceReducer } from './reducer'
+export { pricingServiceEpic } from './epics'
